board:2091 the analysis tail is empty, and a mesher that could not build says so
`make lint` 330 -> 264. What remains is `readability-function-cognitive-complexity`
(145) and `bugprone-easily-swappable-parameters` (119); every other check reads
zero. Twelve of the thirty-two were judgement rather than a fixit.

What Unreal does, what RAGE does: both make the mesher's refusal part of its
signature -- Unreal's `IMeshBuilderModule::BuildMesh` returns bool and the caller
skips the LOD, RAGE's `grcVertexBuffer` allocation refuses and the drawable is
dropped. Neither raises a `noexcept` that terminates. Taken: `StructureMesher::Mesh`
returns whether anything was raised, and both call sites act on it.

`StackProbe` now walks `std::byte *` rather than `uintptr_t`. Painting the stack
needs address arithmetic but never an integer-to-pointer cast: the alignment goes
through `reinterpret_cast<uintptr_t>(top) & 7` -- pointer to integer, which is not
what pessimises -- and the walk stays a pointer.

`ConfigureTemporalResolve` is deleted. The stage never had anything to configure;
`.Encode = nullptr` was already admissible in the executor table and `.Configure`
is now too. This is the pointer-to-member the earlier fixit broke: the finding was
right, the suggested repair was not.

Also: `std::strerror` -> `std::system_category().message` (thread-safe);
`Instancing::Add` uses the refusal channel it already had; `Paves` drops the
unnamed parameter nothing read; `1LL << 62` -> a named `kExponentSpan`;
two `static_cast`-through-void -> `reinterpret_cast`; four file-local functions in
`Declaring.cpp` and one in `TreeGrower.cpp` into anonymous namespaces.

Measured: OldTown 5243fc7d, Venice 50b5f6ae -- both unmoved. Suite 9 PASS,
1 UNPREPARED (CentralPark, unchanged: 3.67 M building triangles built and none in
the frame -- board:2092's, not this commit's).

+++ src/generators/draw/TreeLeaf.cpp
@@ -192,7 +192,9 @@ void BuildPalmate(Sink &sink, const TreeSpecies::Leaf &p) {
       vv[static_cast<size_t>(idx)] = static_cast<float>(i) / static_cast<float>(r);
     }
   }
-  const auto at = [a](int i, int j) { return static_cast<size_t>(1 + (i - 1) * (a + 1) + j); };
+  const auto at = [a](int i, int j) {
+    return 1u + static_cast<size_t>(i - 1) * static_cast<size_t>(a + 1) + static_cast<size_t>(j);
+  };
   for (int j = 0; j < a; ++j) {
     const size_t b = at(1, j);
     const size_t c = at(1, j + 1);
